requirements.txt
anyio==4.0.0
APScheduler==3.6.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
decorator==5.1.1
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.3.0
pycparser==2.21
pytest==7.4.2
python-dotenv==1.0.0
python-telegram-bot==13.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tornado==6.3.3
tzlocal==5.0.1
urllib3==2.0.4
